requirements.txt
numpy==1.26.4
scipy==1.16.2
scikit-learn==1.7.2

tqdm==4.66.5
matplotlib==3.9.2

PyYAML==6.0.2
pandas==3.0.0
omegaconf==2.3.0

tqdm==4.66.5
zuko==1.3.0
torch==2.5.0
torchdiffeq==0.2.4
torchvision==0.20.0
torchdyn>=1.0.6   # 1.0.4 is broken on pypi
torchsde==0.2.6
timm==1.0.12

# conf
wandb==0.16.6
aiobotocore==2.5.4
boto3==1.28.17
dvc[s3]==3.59.0

# multi-run
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
hydra-submitit-launcher==1.2.0
